Address -Wsign-compare

+++ src/battery/battery.cc
@@ -342,7 +342,8 @@ bool Battery::Resize() {
     int new_size = std::max(bat_percentage_width, bat_time_width) +
                    (2 * padding_x_lr_) + (2 * bg_.border().width());
 
-    if (new_size > width_ || new_size < (width_ - 2)) {
+    if (new_size > static_cast<int>(width_) ||
+        new_size < static_cast<int>(width_) - 2) {
       // we try to limit the number of resize
       width_ = new_size;
       bat1_posy = (height_ - bat_percentage_height - bat_time_height) / 2;
@@ -353,7 +354,8 @@ bool Battery::Resize() {
     int new_size = bat_percentage_height + bat_time_height +
                    (2 * (padding_x_lr_ + bg_.border().width()));
 
-    if (new_size > height_ || new_size < (height_ - 2)) {
+    if (new_size > static_cast<int>(height_) ||
+        new_size < static_cast<int>(height_) - 2) {
       height_ = new_size;
       bat1_posy = (height_ - bat_percentage_height - bat_time_height - 2) / 2;
       bat2_posy = (bat1_posy + bat_percentage_height + 2);
